Remove common.js
