--- src/config/prettier-config-explorer/common.js
@@ -1,17 +1,0 @@
-const CONFIG_FILE_NAMES = [
-  "package.json",
-  ".prettierrc",
-  ".prettierrc.json",
-  ".prettierrc.yaml",
-  ".prettierrc.yml",
-  ".prettierrc.json5",
-  ".prettierrc.js",
-  ".prettierrc.mjs",
-  ".prettierrc.cjs",
-  "prettier.config.js",
-  "prettier.config.mjs",
-  "prettier.config.cjs",
-  ".prettierrc.toml",
-];
-
-export { CONFIG_FILE_NAMES };
